Fix mouse cursor going berserk when showing multiple viewports
This was because screen area was associated with the window and not the
viewport as it should.
