--- Source/Engine/AGS.Engine/Graphics/Logic/Interfaces/IAGSWindowInfo.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using AGS.API;
-
-namespace AGS.Engine
-{
-    public interface IAGSWindowInfo : IWindowInfo
-    {
-        new Rectangle ScreenViewport { get; set; }
-    }
-}
