+++ Source/Engine/AGS.Engine/Graphics/Logic/Utils/GLUtils.cs
@@ -51,50 +51,11 @@ namespace AGS.Engine
             _graphics.LoadIdentity();
         }
 
-        public void RefreshViewport(IGameSettings settings, IAGSWindowInfo window, IViewport viewport)
+        public void RefreshViewport(IGameSettings settings, IWindowInfo window, IViewport viewport)
         {
-            float viewX = 0;
-            float viewY = 0;
-            float width = window.GameSubWindow.Width;
-            float height = window.GameSubWindow.Height;
-            if (settings.PreserveAspectRatio) //http://www.david-amador.com/2013/04/opengl-2d-independent-resolution-rendering/
-            {
-                float targetAspectRatio = (float)settings.VirtualResolution.Width / settings.VirtualResolution.Height;
-                Size screen = new Size(window.GameSubWindow.Width, window.GameSubWindow.Height);
-                width = screen.Width;
-                height = width / targetAspectRatio + 0.5f;
-                if (height > screen.Height)
-                {
-                    //It doesn't fit our height, we must switch to pillarbox then
-                    height = screen.Height;
-                    width = height * targetAspectRatio + 0.5f;
-                }
-
-                // set up the new viewport centered in the backbuffer
-                viewX = (screen.Width / 2) - (width / 2);
-                viewY = (screen.Height / 2) - (height / 2);
-            }
-
-            var projectionBox = viewport.ProjectionBox;
-			viewX = viewX + (viewX + width) * projectionBox.X;
-			viewY = viewY + (viewY + height) * projectionBox.Y;
-			var parent = viewport.Parent;
-			if (parent != null)
-			{
-                var boundingBoxes = parent.GetBoundingBoxes(_state.Viewport);
-				if (boundingBoxes != null)
-				{
-                    viewX += boundingBoxes.ViewportBox.MinX;
-                    viewY += boundingBoxes.ViewportBox.MinY;
-				}
-			}
-			int widthInt = (int)Math.Round((float)width * projectionBox.Width);
-			int heightInt = (int)Math.Round((float)height * projectionBox.Height);
-            var viewXInt = (int)Math.Round(viewX + window.GameSubWindow.X);
-            var viewYInt = (int)Math.Round(viewY + window.GameSubWindow.Y);
-            ScreenViewport = new Rectangle(viewXInt, viewYInt, widthInt, heightInt);
-            window.ScreenViewport = ScreenViewport;
-            _graphics.Viewport(viewXInt, viewYInt, widthInt, heightInt);
+            viewport.RefreshScreenArea(settings, window, _state);
+            ScreenViewport = viewport.ScreenArea;
+            _graphics.Viewport(viewport.ScreenArea.X, viewport.ScreenArea.Y, viewport.ScreenArea.Width, viewport.ScreenArea.Height);
         }
 
 		public void GenBuffers()
